+++ lib/DhanHQ/models/multi_order.rb
@@ -26,6 +26,10 @@ module DhanHQ
     #   result.for_sequence("1").order_id
     #
     class MultiOrder < BaseModel
+      extend DhanHQ::Concerns::BangWrites
+
+      bang_class_writes :place
+
       HTTP_PATH = "/v2/alerts/multi/orders"
 
       attributes :orders