+++ lib/DhanHQ/models/twap_order.rb
@@ -45,6 +45,11 @@ module DhanHQ
     #   order.cancel
     #
     class TwapOrder < BaseModel
+      extend DhanHQ::Concerns::BangWrites
+
+      bang_class_writes :create
+      bang_writes :modify, :cancel
+
       include Concerns::ApiResponseHandler
 
       attributes :dhan_client_id, :order_id, :correlation_id, :order_status,